[parser] simplify keyword matching
Summary:
instead of matching an arbitrary identifier, which is a complex regex,
and then checking to see if it's a keyword, just match keywords directly.

also, keywords must be that literal byte sequence, no escape codes are allowed.
identifiers can include unicode and escape codes, so that codepath will get more
complicated and getting keyword handling out of the way will help with
complexity.

Reviewed By: samwgoldman

Differential Revision: D5717860

fbshipit-source-id: e4e002ad3e4fc651785648018534965050ab5460