Fix static super types in classes
Summary:
Before this change, we always treated the `super` field of an instance's
`statics` as the `ClassT super` where `super` is the instance's `super` field.
This was wrong, but in a pretty nuanced way.

When we look up fields on an instance, we walk the super chain. We also insist
that fields declared on the instance be comaptible with fields declared along
it's super chain.

By modeling the "static super" as `ClassT super`, Flow does not accept accessing
Function.prototype fields from class statics, when those fields should be
reachable.

Also, and more relevant to my current efforts, the current modeling makes it
hacky to represent the prototype of `Object` as `null` (Which is accurate.
Observe that `Object.prototype.__proto__ === null`), because we would model the
static super type as `ClassT NullT`, which is nonsense. With these changes, I
can ensure the instance super is null while the static super is still FunProtoT.

Note that this is still not exactly right. In particular, the current model
confuses the notions of "super" and "proto." In reality, a class definition
simultaneously defines an instance type and it's prototype -- fields live on the
instance while methods are defined on the prototype. Subclass's prototype
inherits from the superclass's prototype. The only reason we should walk the
"super" chain is to find these prototypes.

I plan to revisit this to move the internal modeling closer to the runtime
semantics, but for now this change is sufficient to unblock me.

Reviewed By: avikchaudhuri

Differential Revision: D5711631

fbshipit-source-id: b531cd22f82e142026b8c2acfe8828bb9022dc8f