Reimplement find-refs
Summary:
This implements find-refs for local variables using the scope builder, rather
than hooks during type inference. It is more limited than the previous
implementation, as it doesn't work with types, destructuring, and in other
scenarios that the previous version supported. However, it in the limited case
that it does support, the new version yields much more usable results.

Reviewed By: avikchaudhuri

Differential Revision: D5660864

fbshipit-source-id: 7a1b55c39817ddbfa9c27f30babc070689cca5c1