[parser] eliminate mutual recursion
Summary:
I ran into what I think is a js_of_ocaml bug, where it would create
code roughly like this:

```
function recursive_func() {
  for (;;) {
    var x = function(z) { z() },
        y = x(z),
        z = function() {};
    // ...
  }
}
```

which means that when `x(z)` is called, `z` is undefined and everything breaks.
it needs to swap `y` and `z`. The simple recursive function example only repros
with `--disable optcall`, so the tailcall optimization avoids the issue. Since
jsoo doesn't optimize mutual recursion, the `optcall` optimizer doesn't help in
that case.

So, to avoid the issue, this diff rearranges everything to eliminate mutual
recursion. Self-recursion is fine as long as they're tailcalls.

In a couple of places, I made the functions not recursive, and then wrote a
small recursive helper that calls them until they return a `Token` instead of a
`Continue`. I think some of the calls are not tail calls when those functions
are directly recursive, which may be something we can come back and fix later.

Reviewed By: avikchaudhuri

Differential Revision: D5723537

fbshipit-source-id: 55e0d3f92c1712c6489d07757905318252f8d924