Only stream parse errors if there are any
Summary:
During recheck, we stream errors to the persistent connection.
We do this during merge, as errors pop up. We also send all the parse
errors.

However, sometimes there are no parse errors. So we're sending an empty
notification to the persistent connection. If this notification was
interpreted as "these are all the errors", then it would be fine. But
since this notification means "Here are some new errors I found during
the recheck", there's no point in sending an empty notification.

Reviewed By: samwgoldman, nmote

Differential Revision: D5726113

fbshipit-source-id: 3c3ab729754aeff18414c127944426ade5bf19ce