[hack] Delete unused Server_initializing exceptions and exit code
Summary:
The error code `Exit_status.Server_initializing` is never used, so I deleted it. This was introduced in D2194895 and used as an exit code out of clientConnect.connect in case the server was busy. But we have since moved to a monitor which is (just about!) always responsive, so clientConnect itself will never encounter an initializing server. That only comes later, and is reported differently.

The exceptions `ServerMonitorUtils.Server_shutting_down` and `ServerMonitorUtils.Last_server_died` are never used, so I delete them. The exceptions were declared in their current location in D2766909 and used by MonitorConnection.consume_prehandoff_messages, but that function has since been changed to return a Result.t that returns a ServerMonitorUtils.connection_error struct.

Reviewed By: alexchow

Differential Revision: D5570423

fbshipit-source-id: 3b6fbbb9197d89adb97c83efa4a3d2538c4628aa