Ensure opaque IDs are stable during context optimization
Summary:
Opaque types are equated by their IDs, but IDs are generated fresh during
inference. When we recheck, these IDs are re-generated. However, merge
optimizations mean we won't recheck dependents of modules that haven't
"changed" which can cause different modules to disagree about the ID of a given
opaque type.

This diff ensures that opaque IDs are "stable" and will not change spuriously.
However, we have to be careful not to relativize IDs that come from other
modules. For both of these problems, we can re-use existing tricks for nominal
class IDs.

Reviewed By: calebmer

Differential Revision: D5750330

fbshipit-source-id: 6f43200f9dfe0cbaee140181dd66ac2c0cbb006f