Ensure PolyT IDs are stable during context optimization
Summary:
Same issue as D5750330, except this time we might not find errors where we
expect to during rechecks. The "worst case" scenario when PolyT IDs differ is
that we do expensive structural checks. However, this is worse than it might
seem at first. If someone is relying on phantom types, we will miss expected
errors during recheck. If a phantom-type error is suppressed in the code base,
we will show a missing annotation warning, which would be confusing.

Reviewed By: calebmer

Differential Revision: D5752616

fbshipit-source-id: 73d94d6633af1659f43fa1859cd718b88c863a41