Memoize collate_errors
Summary:
On larger repos, `collate_errors` can take a couple of seconds.
By memoizing it, `flow status` goes from taking a couple of seconds to
taking like 40ms.

The one tricky bit is that in lazy mode, sometimes running
`typecheck_contents` causes the checked set to grow. That means it is
not sufficient to regenerate the collated errors only after init or
recheck. When this happens, we'll regenerate the collated errors when
they're requested, which will likely cause the server to appear stuck to
any clients until it is done (which is no better than the current
behavior).

Reviewed By: samwgoldman

Differential Revision: D5755453

fbshipit-source-id: 0c9c734ce6168646b6d43cb177785d357dd34492