Resolve annots in eval_destructor
Summary:
Description:
The code in eval_destructor attempts to preserve unions by recursively creating
EvalTs for each member of the union. However, we might also encounter an AnnotT
which resolves to a union, which should also hit that code path.

However, an AnnotT contains a tvar which might not be resolved yet. We know,
because annotations are 0->1, that it will eventually resolve to a single type.
We exploit the same assurance in speculative match to resolve annotations. This
code uses the same full type resolution strategy as speculative match, but
instead of triggering a "try flow" constraint, we trigger a (new) "eval
destructor" constraint.

When we trigger, the annotation tvar will have been resolved, so we can create
a recursive EvalT with a concrete type.

Reviewed By: avikchaudhuri

Differential Revision: D5740587

fbshipit-source-id: 57327b94e7423e79e459356c179f14a639fd3bc8