use TypeArgCompatibility use_op on ArrT ~> ArrT
Summary:
`ArrT` is a special case of a `PolyT` on the `Array` class. `ArrT ~> ArrT`
is basically `Array<T> ~> Array<U>`, and we flow `T ~> U`. we currently don't
pass a use_op for this, which leads to bad errors that point at `T` and `U`
rather than the `Array<T> ~> Array<U>` where the conflict occurred.

this diff uses `TypeArgCompatibility` on a param called `T`, since arrays are
defined as `class Array<T>` in core.js.

Reviewed By: calebmer

Differential Revision: D5760703

fbshipit-source-id: e43e940000a55fed12a46af15b161d9b93cf0f23