Recursively concretize union-like types in AnnotT
Summary:
The following case does not produce an error as expected:

```
type Fn2 = <A, B>(A, B) => A | B;
declare var fn2: Fn2;
const x = fn2((42: number), ('foo': string));
(true: typeof x);
```

https://flow.org/try/#0C4TwDgpgBAYgdgJigXigHgIIBooCEB8AFNngJQr5QZQA+eA3AFAAmEAxgDYCGATtAG68oAM0QAuWIiZsA9nADOwKAA8UIxIUIAWBBLgBXALYAjCD1I5CAcmEyZViYp4BLOAHNSpJoWA99ECVBIGWEVL0YgA

This is because even though `A | B` is wrapped in an `AnnotT` and concretized, its individual members `A` and `B` are not. So when we flow `boolean ~> A | B` we add a lower bound of `boolean` to `A` instead of flowing `boolean ~> number` and `boolean ~> string`.

This diff fixes this by wrapping union-like members in `AnnotT` so that they may be concretized before they are checked.

I need this fix for the implementation of `$Call` which is then needed for the `$ObjMap` fix.

**Note:** This diff includes the tests samwgoldman wrote in D5717999.

Reviewed By: samwgoldman

Differential Revision: D5722202

fbshipit-source-id: 8c8aaadfe69184281dfd3aa8cdbe619486a95eea