Fix lower propagation in $ObjMap
Summary:
The following case before this diff does not error:

```
({
  foo: 'asd',
  bar: 'asd',
}: $ObjMap<
  {
    foo: number,
    bar: number,
  },
  <T>(T) => T,
>);
```

https://flow.org/try/#0BQbwUABBBmD2sC4IHICGBnAJsgNJCARqgE5Jpa5gC+SAJAPIEBWAsqgA4A8+4UUciCADsArgFsCAU2J4+hEklETpsiFVWcAKgD5gmgJQQAvNoia82-QG4wQA

This is because `$ObjMap` created a tvar for `foo` and `bar` where `number ~> tvar`. Since the object we are casting is a literal we would also see `string ~> tvar` which did not hit the `number` constraint and so we would not see an error.

Reviewed By: samwgoldman

Differential Revision: D5646671

fbshipit-source-id: 475f83c095a5d38ec0c6d4f28476c32d7b827edc