[parser] error about elisions after rest elements in ArrayAssignmentPattern
Summary:
`([...x,] = [])` is an error, because `AssignmentRestElement` must be
the last element in the pattern.

to accomplish this, I changed `object_cover` from tracking a list of locs of
`=`'s that would be invalid when the cover is an expression, to `{ if_patt; if_expr }`
where the locs of the `=`s are in `if_expr`, and the locs of any trailing `,`'s
are in `if_patt`.

also, since it's covering both object and array patterns, I renamed
`Object_cover` to `Pattern_cover`.

Reviewed By: samwgoldman

Differential Revision: D5735038

fbshipit-source-id: b6a55f11cbee78e51d699cd8a6f03f3a5d742034