Terminate proto lookups at NullT
Summary:
Currently we assume that ObjProtoT is at the end of any prototype chain. This
isn't true in practice -- Object.create(null) exists, for example.

This is part of a stack of diffs that untangles this long-held assumption about
prototype chains. As a first step, we should recognize that `null` is the true
terminus of chains.

Object.prototype is still the default __proto__ for instances, objects, etc.,
but the __proto__ for Object.prototype is `null`. This diff encodes that by
special-casing the "Object" builtin interface. If we didn't do this, prototype
lookups might get stuck in a loop.

Note that this diff hacks around some existing issues with class statics. In
particular, we derive the class static super as `ClassT` of the instance super
type. This is wrong in a few ways. For one, class statics have function
prototype stuff, like bind, but `ClassT<ObjProtoT>` does not.

In addition to the existing issues, creating a `ClassT<NullT>` static super
would be very bad, because `ClassT<T>` as a lower bound propagates `T` to
`GetStaticsT` and `(NullT, GetStaticsT)` is (rightly) undefined.

This diff includes a very hacky and hopefully temporary workaround, which uses
FunProtoT instead of ClassT<NullT> for only the Object builtin type.

Reviewed By: avikchaudhuri

Differential Revision: D5263417

fbshipit-source-id: bef0227cccd25749538ef1096f3449bdb45734ff