Optionally replace reason desc during annotation reposition
Summary:
This diff adds the capability for repositioning to replace the description in
addition to the location of a reason.

In future diffs, I'm going to use this feature to eliminate changes in error
messages, which are caused by builtin type lookups on FunProtoT and ObjProtoT,
whose reason descriptions are replace from e.g. "arrow function" and "object
literal" to "Funcion" and "Object", which is confusing and unhelpful.

Reviewed By: gabelevi

Differential Revision: D5153455

fbshipit-source-id: 83d294e1664b793c71d3c2d352ce318ada3738be