Delegate FunProtoT lowers to Function.prototype for all uses
Summary:
FunProtoT has some special behavior as a lower bound. Specifically, it
early-terminates property lookup (LookupT). After any of these special cases,
though, we can just treat it like an InstanceT.

This means that FunProtoT can be compatible with ObjT, or anything that
InstanceT is compatible with. Note that we will not be able to evaluate
constraints involving a FunProtoT lower until merge, when the Function instance
type is available.

Reviewed By: calebmer

Differential Revision: D5153470

fbshipit-source-id: 1e2fed9fb00ab141e74779c21cc682894006b39c