Propagate FunProtoT upper bounds to Function.prototype
Summary:
Currently, in `ground_subtype`, we assume that anything is a valid subtype of
`FunProtoT`. That is, we assume that `FunProtoT` is effectively `mixed`, which
is absurd and wrong.

I want to change that rule, but there aren't any rules that specify what should
actually happen when FunProtoT is an upper bound. This diffs adds that rule.

Note that this change broke the constructor_annots test, which was designed to
support subtyping a ES3-style class into an interface. To fix I needed to add
FunProtoT as a static super to interfaces with a static callable (as we do with supers
for non-static callables) and add logic to the `ClassT _, ExtendsT _` rule, which fires
now when it didn't before.

Reviewed By: avikchaudhuri

Differential Revision: D5153486

fbshipit-source-id: 9b2ae640bc88e360b9615d11e99f1719022e9790