Delegate ObjProtoT lowers to Object.prototype for all uses
Summary:
ObjProtoT has some special behavior as a lower bound. Specifically, it
early-terminates property lookup (LookupT). After any of these special cases,
though, we can just treat it like an InstanceT.

This means that ObjProtoT can be compatible with ObjT, or anything that
InstanceT is compatible with. Note that we will not be able to evaluate
constraints involving a ObjProtoT lower until merge, when the Object instance
type is available.

Reviewed By: avikchaudhuri

Differential Revision: D5153525

fbshipit-source-id: d94fbf842f7649809f9b3831415874e319df7d20