Implicit/global `this` type in functions/calls should have a location
Summary:
Functions and function calls have an implicit `this` type. Method calls use the
subject of the method as `this` -- e.g., in `o.m()`, `this` in `m` is `o`.

Currently the implicit `this` type is `ObjProtoT`, which is a weird type,
because it behaves like `mixed` as an upper bound (see `ground_subtype`), but
behaves like the builtin `Object` as a lower bound.

In order to support `null` prototypes, we need to resolve the weirdness of
`ObjProtoT`, which means we need to actually check `L ~> ObjProtoT` constraints
as if the upper bound were the builtin `Object` type.

However, doing so causes locationless errors when `this` types are
incompatible. These errors are legit, but without a location the error output
is unactionable.

Reviewed By: mroch

Differential Revision: D5756629

fbshipit-source-id: 65a515171d29dbbe8ed8743532ff06dcbc5d1c5b