Change Class_sig.field repr
Summary:
In a later diff, it turns out to be fairly convenient to keep the Func_sig
representation of field initializers until later, which is what this diff does.
Either a field has an initializer and is a `Func_sig.t` or it doesn't and
it's just some `Type.t` -- the annotation or a tvar.

Reviewed By: avikchaudhuri

Differential Revision: D5609594

fbshipit-source-id: 3de0784abbe03e314d024d9ed482db41308dd44a