Reposition upper bound object prototype when checking object lower bounds
Summary:
Currently, ObjProtoT behaves like mixed as an upper bound. I want to fix this,
which means that the often locationless ObjProtoT now might appear in error
messages. We can work around this by simply repositioning the proto object to
the location of the original upper bound object, which will certainly have a
location.

Reviewed By: avikchaudhuri

Differential Revision: D5759767

fbshipit-source-id: 631cc9516779da471176df235446a8a9e5de0688