Propagate ObjProtoT upper bounds to Object.prototype
Summary:
Now that we merge the master cx before local inference, we can forward all
constraints involving ObjProtoT to the builtins.

Before, we avoided this, because doing so caused lots of constraints to become
deferred until the end of merge, which was bad for performance but also opened
the door for any-unification to reduce coverage.

Reviewed By: avikchaudhuri

Differential Revision: D5681521

fbshipit-source-id: d6ff24ba632c0b7ee011c5d26979e70e1d39e876