Support __proto__ property for object initializers
Summary:
An object literal can specify a `__proto__` property, which is does not create a
normal property but rather creates an object with the specified value as its
prototype instead of Object.prototype.

This was non-standard behavior, implemented by all browsers (and Node), but it
was recently standardized in ECMAScript 2015 as an optional, normative
specification.

http://www.ecma-international.org/ecma-262/6.0/#sec-__proto__-property-names-in-object-initializers

Reviewed By: gabelevi

Differential Revision: D5121709

fbshipit-source-id: 80f4d278c084a7c81937e395bf017d631f1b8715