Support __proto__ property for object type annotations
Summary:
In order to be more precise about property layout, it's useful to be able to
specify that layout in an annotation. Before this diff, any object type always
had the same prototype: Object.prototype. (Function.prototype for callable
objects).

This diff makes it possible to specify prototype chains in an annotation. Note
that the existing object subtyping and property load/store machinery already
supports arbitrary prototype chains.

There is a bit of follow-on work, which I will do separately:

1. Improve error messages. Like property compatibility errors, these prototype
   checks become nested, and the deeply-nested positions are potentially far
   away from the original use.

2. Assume that all other properties are "own" when a __proto__ property is
   present. This can be useful for spread, where properties which might be
   non-own become optional. (Note that width subtyping would still apply, so
   non-overlapping properties would still become mixed in the spread result.)

3. Detect and error on loops. Currently loops are pruned by the constraint
   cache, but ideally we could error on loops, because such a prototype chain is
   actually malformed.

Reviewed By: gabelevi

Differential Revision: D5121953

fbshipit-source-id: d68927e99f7dcf81bf6cdd2dabd5c732348d70a6