Validate prototype values are objects
Summary:
Setting a prototype, using Object.create, __proto__ properties, and the
unsupported Object.setPrototypeOf requires that the value be an object (or
null). Otherwise, it silently uses Object.prototype.

This diff partially implements this behavior. If the value is not object like,
we will emit an error, which is stricter than runtime but feels like the right
thing to do, and fallback to Object.prototype.

For __proto__ properties in annotations, I needed to use a type destructor
because the __proto__ field can appear in upper bound positions. This is the
same ReposUseT "slingshot" trick that all the other type destructors rely on.

Will add support for null prototype objects in a follow-up commit.

Reviewed By: mroch

Differential Revision: D5124564

fbshipit-source-id: 686b90a3f2c18b6b7eedebb3e0de3476567e6a54