get rid of special scope handling for catch clauses
Summary:
We had bent over backwards really hard to ensure that when a variable declared
inside a catch clause has the same name as the catch parameter, the two names
are kept in sync. Otherwise, what might happen is that the assignment
corresponding to the variable declaration would erroneously affect the catch
parameter and not the actually hoisted variable outside the catch clause. Or
worse.

However, with support for preserving variable name shadowing now properly in
place, we can remove the shady logic for this special case.

(Note: this ignores all push blocking failures!)

Reviewed By: mroch

Differential Revision: D5790833

fbshipit-source-id: 7aebd005ab192cf92c2337bc72cc1750f035eab5