Better support for tvars in type destructors
Summary:
In the current implementation of type destructors if we are destructuring a tvar then we don't concretize the result. This is because we can't be certain that the tvar we are destructuring will resolve at all, or if it will resolve that it will only resolve once. Before it was assumed that tvars could never reach a type destructor, however in hindsight polymorphic function type variables could easily pass a tvar into a type destructor. Here is an example of where bugs could creep in:

```
class C<T> {x: T}

declare function fn<T>(x: C<T>): C<$PropertyType<T, 'foo'>>;

declare var c: C<{foo: {bar: any}}>;
const x = fn(c);

(x: C<{bar: mixed}>);
(x: C<{bar: empty}>);
```

You would not expect an error on this example, but when you run Flow against this program you do\! Flow is unifying `{bar: mixed}` with `{bar: empty}` and so you get `mixed ~> empty` errors. This happens because `$PropertyType` is an open tvar so when we flow `C<$PropertyType> ~> C<{bar: mixed}>` we unify `$PropertyType` and `{bar: mixed}` so that `{bar: mixed}` becomes a lower of the `$PropertyType` tvar. When we do the second type cast the `{bar: mixed}` lower is then unified with `{bar: empty}` and we get our unexpected error.

This diff fixes this error by unwrapping resolved tvars. I'm unsure if this is a coincidental solution, or fundamentally the right solution. This works because of variance. We resolve `T` when we unify `C<T>` with `C<{foo: {bar: any}}>`. So when the resolved `T` reaches `$PropertyType` now with this diff we unwrap the resolved type and don't run down the tvar path which returns an open tvar for `$PropertyType`.

So what happens when we don't unify? Take the following example. It is the same as before, but without `C<T>` which means we don't unify and resolve `T`.

```
declare function fn<T>(x: T): $PropertyType<T, 'foo'>;

declare var c: {foo: {bar: any}};
const x = fn(c);

(x: {bar: mixed});
(x: {bar: empty});
```

If you run Flow against this program today, without this diff, you don't get an error\! Unlike in the example above where you would get an error without this diff. We don't get an error because `{bar: mixed}` is only ever an upper bound for the tvar returned by `$PropertyType`. Values may never be used as an upper bound and so the `$PropertyType` tvar will never get any new lowers. (Using `typeof` does allow values as upper bounds, but `typeof` doesn't allow new lowers because it resolves the type.)

So it does seem like there is an argument to be made that this is fundamentally the correct solution to the problem.

Reviewed By: samwgoldman

Differential Revision: D5801885

fbshipit-source-id: 8f63bc2ee0762457c9678f08b813ed7768437149