Remove implicit tvar traversal from type mapper
Summary:
The only place we use the type mapper is in subst, which has an override to
ensure we don't traverse tvar bounds.

This makes the type mapper more like the type visitor, which doesn't traverse
tvars, but allows subclasses to override the tvar behavior.

The previous behavior required a separate Root_finder module to break a cycle
between Type_mapper and Flow_js, which is now unnecessary.

Reviewed By: avikchaudhuri

Differential Revision: D5788917

fbshipit-source-id: 7cffd8e2a54a9462291b967642aa7325964e4dca