Preserve use_op through ExactT upper bounds
Summary:
When a type flows to an ExactT upper bound, we store the lower bound in a
continuation and exactify the wrapped upper bound, then continue.

Previously, we were forgetting the use op from the original constraint to
ExactT. Instead, we can store the op in the continuation along side the lower
bound.

Reviewed By: calebmer

Differential Revision: D5803425

fbshipit-source-id: 02a4d02c8b63ad5829a219ac8d8287db2ea02291