Unwrap use ops in EPropPolarityMismatch error
Summary:
When an inexact lower bound flows into an exact type, we emit an error. If that
constraint has a use op, we would not display it, potentially missing important
context for the error.

Note that the variance incompatibility error still points to the objects, but a
further improvement would be to make it point to the property types, since it
will be nested in a use op that points to the objects.

Reviewed By: calebmer

Differential Revision: D5806440

fbshipit-source-id: 8453acb5780b965147ed6a0cc8187bf15dd457eb