Preserve use_ops through UnionT speculation
Summary:
When a type flows to a UnionT upper bound, we go into a "speculative match"
workflow, and eventually emit constraints against each arm of the union.

Before this change, those eventual constraints "forgot" the original use_op.
This just threads the use_op through to the right place, similar to D5803425.

Reviewed By: calebmer

Differential Revision: D5807267

fbshipit-source-id: 0d4d96c8cd610e994b0e6089f2ae6487a3e1787a