Retool error positioning for property reads/writes
Summary:
Previously, we repositioned property reads consistently, but writes were fairly
inconsistent.

Both reads and writes can be viewed as a single constraint between types with a
source and a sink. For reads, `x = o.p`, we want to reposition the source to
the location of the read. This is trivial with ReposLowerT, and was performed
in various places, like GetPropT, MethodT, and CallT.

For writes, we want to reposition the sink, which is harder. The sink might
just be a tvar, designed to be widened arbitrarily, so no error will ever
occur and repositioning is unnecessary. If we tried to wait for the sink to
resolve, we might stall, so standard repositioning approaches don't work.

Because of this, we ultimately never repositioned writes. There was one
exception, setting a property on an InstanceT. This was managed with a ReposT
upper bound. However! This only worked on writes directly to that InstanceT.
Once we walked the prototype chain, no repositioning happened.

For example:

```
class A { p: number }
class B extends A { q: number }
var b = new B;
b.q = null; // error, number repositioned to b.q
b.p = null; // error, number points to declaration in A
```

ObjT writes were never repositioned.

This change makes it so that all writes are repositioned to the property
reference in the assignment statement, regardless of what they are (instances
or objects) and how they were looked up (directly, or through the prototype
chain).

I made a few changes to support this new behavior:

* Helpers to read/write props, like get_prop, set_prop (instances) and
read_obj_prop, write_obj_prop (objects) now call `perform_lookup_action`
instead of directly emitting constaints. This makes it possible to consolidate
repositioning logic into one place.

* All repositioning in GetPropT, MethodT, CallT, and LookupT has been removed.
If we need to reposition a read or a write, it will happen once the property
has been found.

* Some instances of the RWProp lookup action have been changed to LookupProp
actions. RWProp should only be used if a property is being read syntactically,
as it will be repositioned to the op reason.

* Ops surrounding property writes have been removed. Because we didn't
reposition writes to the property reference location, we used this op to help
localize errors. The op is now always redundant, where it was just sometimes
redundant before.

There are a couple warts in this design, which are not too significant, but
worth calling out.

* Writes are only repositioned if the op reason is RPropertyAssignment. An
earlier version of this change always repositioned, but that caused regressions
in legit uses of `SetPropT` like `Object.assign`, which should not be
repositioned.

* Removing the op for property writes regresses errors on setters where the
setter function does not have an annotation. This is covered by a test, where I
added a TODO comment. I think the overall improvement to error localization
overshadows this regression.

Reviewed By: mroch

Differential Revision: D5817628

fbshipit-source-id: 1c5506502d5e43ae3ca9429872cc8b15a02c8560