Preserve goto nodes in context opt
Summary:
Flow implements unification using union-find, where a class of equivalent tvars
all point to the same root node.

Context optimization replaces a tvar, which might be a "goto" node, with a root
node resolved to the union of its lower bounds. If a module signature contains
goto nodes, they would be replaced with a root node.

Potentially equivalent nodes would be replaced with copies of the same
underlying type, bloating the size of the sig context. Furthermore,
optimizations to bypass unification on equivalent tvars would be missed
downstream.

Reviewed By: avikchaudhuri

Differential Revision: D5791060

fbshipit-source-id: 1e80dcf8306308c49001293eb47ec8d38a6cbaa5