Allow strict writes to inexact objects
Summary:
Before this change, any write to an inexact object was forbidden, unless that
object had an own property with the given name or the write matched a
dictionary on that object.

This behavior also differs from "instance" types in Flow. In a later diff, I
make a change to model class statics as object types. But we run into this
issue, based on this pattern:

```
class A {
  static p: number;
}

class B extends A {
}

B.p = 0; // should be allowed
```

Indeed, for some object `b` with proto `a`, where `a` has some property `p`, as
long as we allow `b <: a`, and we allow `a.p = x`, then we effectively also
allow `b.p = x` through subtyping.

For example, it wouldn't be effective to forbid `B.p = 0` above, because I
could just write this:

```
var A_actually_b: Class<A> = B; // allowed by existing rules
A_actually_b.p = 0; // allowed by existing rules
```

This diff makes objects behave like instances, meaning that we allow writes
to objects if the written-to property appears in the prototype chain. To be clear,
this does not allow "unsealed writes" which could add arbitrary properties. I'm
calling this kind of write a "strict write," borrowing language from `LookupT`.

Note that this does break the symmetry between "sealed" in objects and the
semantics of `Object.sealed` at runtime. However, note that the symmetry is
already broken through subtyping (see above), so if we wanted to model
runtime-sealedness, we should add subtype restrictions similar to those for
exact objects.

Still, maybe we should use a different name in our flags to avoid further
confusion. I'm open to suggestions.

Reviewed By: avikchaudhuri

Differential Revision: D5819897

fbshipit-source-id: e5f137ef7f82b001814db478c3ffcf41e5cc4f93