remove most libs from Try Flow
Summary:
these libs are causing a stack overflow. this is a general problem with
the way Flow_js.flow recursively resolves constraints, since resolving one
constraint can cause many other constraints to be unblocked. in a worse-case
scenario, the call depth could be O(number of constraints).

this isn't a problem in the native build, but some combination of
js_of_ocaml and V8's stack size causes problems. while we work on fixing that,
turning off the libs temporarily unblocks us.

it also makes Try Flow initialize super quickly. so in the future, I think we
should only add these back on demand, via checkboxes in the UI.

Reviewed By: avikchaudhuri

Differential Revision: D5822877

fbshipit-source-id: 5ff05da48cad8b62c43e27e2aff94f2d0112b4e6