Add recursion check to __unify
Summary:
I was recently debugging an error where Flow was not terminating. This is because we were recursing in `__unify` and exceeding our trace limit, but because we do not have a recursion check we weren't terminating.

This should make future debugging of non-termination issues that operate only within `__unify` easier.

Reviewed By: mroch

Differential Revision: D5818491

fbshipit-source-id: a1a9c20de5a7e57f5e9860d05a6828b896f464ed