Don't allow class statics to pun methods and fields
Summary:
Unlike instances, where methods live on the prototype and fields are own, class
statics are all always own. Declaring a field after a method should overwrite
the method.

This is prelude to future work, which will model class statics as an object
type instead of an instance type. It's useful to maintain an invariant that the
fields/methods/getters/setters maps in the static sig are disjoint, so we can
coalesce them into a single prop map.

Reviewed By: avikchaudhuri

Differential Revision: D5817862

fbshipit-source-id: 642eec4efe8bb64065af4056cd1389b6bbd2e525