Represent class statics as an object type, not an instance type
Summary:
In a class declaration, methods are defined on the prototype object and fields
become own properties of instances. We model this in our InstanceT type as
separate property maps for fields and methods.

We also use an InstanceT to represent statics, but statics do not share the
same behavior as instances described above. Methods and fields all become
simple own properties.

Furthermore, InstanceT has a "super" field while ObjT has a "proto" field. The
super field on InstanceT is actually pretty confusing, because we treat it like
proto, but it's not. The proto of an instance is actually an object with all
the methods in methods_tmap.

My ultimate goal is to clean up the super/proto confusion in instances, but
first I need to eliminate the use of InstanceT to describe statics.

Reviewed By: avikchaudhuri

Differential Revision: D5819896

fbshipit-source-id: eb04ab9e3df180e416dcc02a31f5507e53430840