`--lazy-mode ide`
Summary:
`flow server --lazy` is the same as `flow server --lazy-mode=fs`.
It separates checked files into focused files, dependents, and
dependencies. Any file that changes on the filesystem is a focused file.
Any file that depends directly or transitively on a focused file is a
dependent file. Any direct or transitive dependency of a focused file or
a dependent file is a dependency file.

`flow server --lazy-mode=ide` adds a file to the focused file set when
it gets a `didOpen` notification from a persisitent connection. At the
moment, `didClose` does not shrink the set (this can be tricky).

The one small caveat to both lazy modes is that running a
`typecheck_contents` against `foo.js` will add all of its direct or
transitive dependencies to the set of dependencies (unless they're
already focused or dependents).

Reviewed By: avikchaudhuri

Differential Revision: D5798472

fbshipit-source-id: 2fe1990eaa6a4e93a6bca12a6f70836b3dffcbad